Add working multiply functionality to calculator!
WOW, a major breakthrough!

When I first started this project, I was not expecting to hit such a wall early on, indeed the hard part so far was figuring out a way to be able to store the 2 sets of numbers (a,b)!

It's easy to store the initial display value in (a) but once you try figuring out how to store them on (b) it gets tricky. I managed to do this by simply keeping count of the click times. I came across the idea of counting the steps necesary to obtain the result. I believe it's 4 step.

First step you input the number you want to operate with, second step is to chose an operator, third step is inputing the new set of numbers and 4th step would be to either press the same operator to obtain the result OR press the result button to obtain the actual result!