Add working calculator
Alright folks! This project took many days mainly because I had to take care of some IRL stuff but the core logic of it was done in about 3-4 days. Like I stated in the earlier commits, the main logic behind the program is using steps defined by the user clicks to determine the state in which the calculator is. Also assigned a type for each operator, this way the calculator which operator was last pressed when obtaining a result.

Overall an awesome project and without a doubt the hardest so far!